dom/nodes/Document-constructor.html

<!doctype html>
<meta charset=utf-8>
<title>Document constructor</title>
<link rel=help href="http://dom.spec.whatwg.org/#dom-document">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var doc = new Document();
  assert_true(doc instanceof Node, "Should be a Node");
  assert_true(doc instanceof Document, "Should be a Document");
  assert_false(doc instanceof XMLDocument, "Should not be an XMLDocument");
}, "new Document(): interfaces")

test(function() {
  var doc = new Document();
  assert_equals(doc.firstChild, null, "firstChild");
  assert_equals(doc.lastChild, null, "lastChild");
  assert_equals(doc.doctype, null, "doctype");
  assert_equals(doc.documentElement, null, "documentElement");
  assert_array_equals(doc.childNodes, [], "childNodes");
}, "new Document(): children")

test(function() {
  var doc = new Document();
  assert_equals(doc.URL, "about:blank");
  assert_equals(doc.documentURI, "about:blank");
  assert_equals(doc.compatMode, "CSS1Compat");
  // XXX Spec says "utf-8", browsers do "UTF-8".
  assert_equals(doc.characterSet.toUpperCase(), "UTF-8");
  assert_equals(doc.contentType, "application/xml");
}, "new Document(): metadata")
</script>
